requirements.txt
python-dotenv~=0.3.0
fastapi~=0.112.2
uvicorn~=0.30.6
SQLAlchemy~=2.0.32
pydantic~=2.8.2
jinja2
requests~=2.32.3